correct operators
When processing dates  `message_udate($tstamp)` should show the hour and minute for the current day and month and day for messages older than the current day. The comment line reflects this, however the wrong operator was used. Alternatively, it could just be `=` rather than `<=` because the current day should never be less than message day.